fix(ci): move telemetry callbacks to client component
